ci: tweak test step
To avoid a `Makefile` dependency issue, a previous commit changed the
location of the download directory to live inside the build directory;
this change propagates that to the CI configuration. Also, it is no
longer necessary to clean up anything between runs: both the `build` and
`run` directory are subdivided by target triple so repeated builds will
not interfere.