[ci] Pin Windows runner to Windows 2025 (#566)
Rather than using Windows-latest this PR updates the ci to use a pinned
version of Windows (set to the latest version available). This is so if
you happen to have an issue with the Windows build you know exactly
which version of Windows it is running, rather than having to dig into
the ci, or Github runner documentation.