Pin MacOS Github runner version in ci to MacOS 15 (#565)
Rather than using Macos-latest this makes the ci used a pinned version
of Macos (set to the latest version available). This is so if you happen
to have an issue with the MacOS build you know exactly which version of
MacOS it is running, rather than having to dig into the ci, or Github
runner documentation.