test: add browser test harness (#563)
This commit adds a browser test harness to run the tests in the browser.

## Motivation

We are heavily using wasi-libc on browsers but we don't have any test
for the case in wasi-libc. In theory, there should be no behavioral
difference between wasmtime and browsers (as long as WASI
implementations are correct) but browsers have their own limitations in
their Wasm engines. For example, memory.atomic.wait32 is not permitted
on the main thread.

We are working on adding some mitigation for such browser-specific
issues (https://github.com/WebAssembly/wasi-libc/pull/562) and this test
harness will help to validate the fix.