Preserve file timestamps when copying files in the install target
This change makes it so that the timestamps of installed header files are
only updated when the contents of the header files change. This change is
beneficial for incremental builds of projects depending on wasi-libc. For
example, the Swift project builds wasi-libc as part of its build process,
and this change reduces the number of build products in the Swift build
that are unnecessarily rebuilt.