Refactor how compiler-rt is managed in CI and locally (#591)
Building the PIC version of libc.a is currently broken with LLVM 19+,
and it's due to how this repository handles the compiler-rt library.
This PR refactors the handling to automatically download an appropriate
copy and handle it internally or otherwise use a `BUILTINS_LIB` from the
environment. This means that it's now always available for `libc.so`
which is needed to fix the build issue with LLVM 19+