Optional SIMD memrchr (#598)
Continuing #580. `memrchr` is mostly used to implement `strrchr`, so
that's what I test.

It has the advantage that we know the length and can access the entire
buffer, so no undefined behavior for this one. I does do unaligned
reads, however, so `wasm_v128_load` is used to dereference the pointer.
It uses SIMD while there are 16 or more bytes to read, then fallback to
scalar.

The only other notable feature is using `clz` rather than `ctz`.