Optional SIMD memcmp (#603)
This the last PR I'm putting forward for #580.

`memcmp` has the advantage that we know the lengths and can access the
entire buffers (no undefined behavior). It has the difficulty that the
buffers may not share an alingment, so it uses `wasm_v128_load`.

It uses SIMD if there are 16 or more bytes to read, otherwise it
fallbacks to scalar. If the number of bytes is larger than 16, but not a
multiple of 16, the second iteration retests some already tested bytes,
to "align" the remaining length to a multiple of 16. Making the first
(rather than the last) iteration special unnecessarily "wastes" these
comparisons, but helps the compiler partially unroll the loop.