Disable build symbol check by default
I've personally found this to mostly be an annoyance having it
on-by-default. This means that changing flags can often break the build
and it also means that externally using a different LLVM can often break
the build too. In the interest of making the build less brittle this
disables the check by default.

To avoid losing the benefit of having this check, however, CI opts-in to
enabling this check for all tests. That should mean that we're still
testing in this repository, but integrations elsewhere don't have to
worry about it.