Add sockets tests for hangups on both sides, and implement timeouts for TCP sockets (#636)
The new tests are based on the previous simple test where the server
accepts one connection and echoes back a message received from the
client.

Scenarios tested:

* Client hangs up after connecting
* Client hangs up while sending
* Client hangs up after sending
* Client hangs up while receiving
* Server hangs up after connecting but before receiving data from client
* Server hangs up while receiving data
* Server hangs up after receiving all data but before sending data back
to client
* Server hangs up while sending data back to client

As part of this testing, I implemented timeouts for TCP sockets. There
are no tests for UDP sockets yet, so timeouts are not implemented yet
for them.