Reimplement directory streams
The main goal of this commit is to simplify what's located in the
descriptor table of libc. Primarily directory streams are no longer
located in the table and are instead stored directly in the `DIR*`
structure. Additionally the wasip1-style `cookie` is removed in favor of
a wasip2-specific offset counter. Along the way
`__wasilibc_nocwd_scandirat` was rewritten for both wasip1 and wasip2 to
be built on `openat` + `fdopendir`. This removes WASI-version-specific
code and makes this look more similar to `scandir` elsewhere in libc.