wasip2: Improve runtime of sockets-related tests (#653)
Don't unconditionally sleep for a second during tests to wait for a
server to come online. Given that tests also are required to run
serially this causes tests to take quite a long time to run. Instead
bake into tests the ability to do a sort sleep but otherwise try to
connect in a loop.