Improve stdio lazy init (#650)
Ensure it's always initialized in descriptor table functions instead of
only having initialization in some functions. Fixes `fstat`, for
example, on 0/1/2.

The previous initialization functions documented that the `file_handle`
wasn't ever touched so it could stay uninitialized, but this isn't true
given other functions throughout wasi-libc. Eventually I'd like to have
a custom descriptor type just for stdio so stdio doesn't have a "null"
`file_handle` field but for now that's the easiest way to slot this in.
While I'm getting other tests working in Python this is in theory enough
to get up off the ground.

More functions will need to grow tests for `__handle == 0` to handle
stdio in the future, but I plan on handling them as they arise.