wasip2: Don't reimplement `preopens.c` (#655)
This is really tricky path-related code and it's currently buggy on
wasip2. Instead of duplicating it like before instead have only a single
definition with some small handling of wasip1-vs-wasip2. This fixes
`chdir`, for example, and handling of preopens. The fundamental bug from
before is that a `wasip2_string_t` was handed out as a `const char*`,
but it wasn't null-terminated. The bug is more apparent with all the
code paths interleaved here.

This additionally fixes an issue where the previous implementation would
continuously open new file descriptors for preopens. Instead now a
single file descriptor is registered at preopens-registration-time which
is then used throughout the lifetime of the program.