wasip2: Merge FILE_{HANDLE,STREAM} types (#656)
This commit merges two different kinds of files in the descriptor table
into one entry. This reduces the number of possible states to manage and
notably fixes bugs where a "stream" is needed but a "file" is present.
Previously the `read` and `write` syscalls would, for example, lazily
transition a `FILE_HANDLE` to a `FILE_STREAM`. This logic was
duplicated, however, and wasn't applied to other functions like `lseek`.
This meant that it was not possible to `open` a file and immediately
`lseek` it as that would erroneously fail.

This commit merges the two types together and transitions all
streams/pollables in the new type to be lazily initialized. This avoids
eagerly creating them, like before, and enables gracefully creating them
as necessary. Along the way the `isatty` function was drastically
trimmed down since it needed changes anyway and the previous version was
returning "true" when creation of a stream returned `ENOTTY`, which
didn't feel right anyway. This implementation will need updating for
stdio streams anyway in the future.

Another change made here is that the descriptor table now explicitly
holds `*_own_t` types instead of `*_borrow_t`. This doesn't actually
change anything at runtime but more accurately reflects the ownership
semantics.