wasip2: Update descriptor table apis (#660)
This is a series of commits, each of which passes tests on its own,
which changes the API of the `descriptor_table_*` functions. The main
focus here is around setting `errno` automatically to handle a number of
possible error conditions and then additionally slimming down the API
slightly -- removing `descriptor_table_update` which is a bit confusing
in the face of `descriptor_table_insert`.