Add gethostname.c to the build (#664)
This is mentioned as not being present in #196 with some discussion
about what to do about it ranging from removing the header to adding a
polyfill. The file iteslf is a small shim around `uname` which is
already included and has a stub implementation. Given that it seems
reasonable to go ahead and include this too as questions about whether
or not it should be present are likely equally applicable to `uname`.

Closes #196