wasip2: Use virtual dispatch in the descriptor table (#662)
This commit is a large refactor of the internal representation of the
descriptor table. The high-level change happening here is that the
tagged-union approach of descriptors from before is now replaced with a
virtual-dispatch approach. The rationale for this change is:

* With a tagged-union approach it means that all functions everywhere
must handle all possible descriptors. This has notable a notable effect
on the output binary meaning, for example, that if you `close` a file
descriptor then `wasi:filesystem`, `wasi:cli`, and `wasi:sockets` are
all imported. This is because `close` needs to know how to close all
kinds of descriptors and it doesn't know ahead of time which it is. With
a vtable-style approach it means that `wasi:sockets` isn't imported
unless you actually create a socket.

* Code-organization-wise the handling of various descriptor types is
pretty sprawling throughout the codebase. This means that changing or
updating fields in descriptors requires consulting all functions that
operate on descriptors which is a tedious and error-prone task. With a
virtual dispatch approach the entire implementation is in a single file
and can be more easily audited.

* Adding/removing descriptor types is a tedious and error-prone task.
All `switch` statements of a descriptor tag need be consulted when doing
this. With a vtable-style approach adding or removing a descriptor type
is trivial as it's just a deletion/addition of a single file.

* This opens up the door to more easily customizing certain kinds of
descriptors. For example this fixes previous issues related to the fact
that stdio is modeled as a file without a handle but with streams. Stdio
now officially isn't a file which means that all file operations fail on
it and don't need to explicitly check for `__handle == 0`. Furthermore
it should be much easier to implement `isatty`, for example, with stdio
descriptors without having to update everything else too.

* This helps deduplicate internal implementations of some functions. For
example the `poll` function is greatly simplified for WASIp2 here.
Additionally the `fcntl` function I plan on updating to have more
support for various operations and won't have to duplicate various
option-handling across various descriptors now.

* This makes it easy to fix a previous issue where `read` and `write`
didn't work on TCP sockets. Now by implementing the various low-level
vtable functions this is trivially supported.

Overall this is a lot of code movement. I additionally did not limit
myself to just code movement and took the liberty to reformat code to a
consistent style as well as auditing/cleaning up existing
implementations. The structure is intended to all be the same with no
foundational change except for the tagged-union-to-vtable change,
however.

While not intentional, this change ended up fixing a number of tests in
Python when building against a WASIp2-version of wasi-libc.

Closes #521