Improve build reproducibility of stub archives (#682)
Currently, the `default` target produces byte-identical binary
artifacts on Linux and macOS, except for the stub archives (e.g.
`libcrypt.a`).

This is because `llvm-ar` defaults to `--format=darwin` on macOS, which
generates an empty symbol index for an empty archive.

We can make this build completely reproducible across macOS and Linux by
specifying `--format=gnu`, which is the default used on Linux.