Optional SIMD str(c)spn (#597)
Continuing #580, implements `strspn` and `strcspn`.

This one follows the same general structure as #586, #592 and #594, but
uses a somewhat more complicated algorithm, described
[here](http://0x80.pl/notesen/2018-10-18-simd-byte-lookup.html).

I used the Geoff Langdale alternative implementation (the tweet as since
disappeared) which is correctly described there but has a subtle bug in
the implementation:
https://github.com/WojciechMula/simd-byte-lookup/issues/2

Since the complexity needed for `__wasm_v128_bitmap256_t` is shared for
both `strspn` and `strcspn`, I moved the implementation to a common
file, when SIMD is used.

The tests follow a similar structure as the previous ones, and cover the
bug, which I was found through fuzzing.