Sever wasip2's dependence on wasip1 headers
This commit updates `wasi/api.h` to be fully undefined on WASIp2
targets. This prevents leaking WASIp1 definitions on unapplicable
targets, such as `wasm32-wasip2` and the upcoming `wasm32-wasip3`.

This involved changing a number of `#define`s from something symbolic to
a number literal. Static assertions throughout the codebase, already
present, assert that these numbers match WASIp1 ABI values and for
WASIp2+ these numbers are just a libc abstraction, not part of the ABI.