Add `wasi/version.h`
As a step towards #688 this adds a new header which ensures that the
`__wasipN__` variables are defined. These will eventually be provided
via llvm/llvm-project#165345 but in the meantime this uses a header to
define them.

This then additionally refactors the internals to have
`wasi/wasip{1,2}.h` as well as `wasi/api.h` which includes the correct
header. The previous `__wasilibc_use_wasip2` macro is removed entirely
in favor of `__wasip2__` defined in this header.