Enforce clang-format style for files in this repository  (#692)
This commit adds necessary CMake logic to use `clang-format` to format
source code contained in this repository. This notably excludes files
such as musl and other upstream source code which isn't primarily
authored in this repository, so it's intended to only apply to code that
is primarily developed here.

Procedurally this PR is split into two commits: the first is the
CMake-logic necessary to add a `format` and `format-check` target. The
second is the actual formatting applied to this repository. I'll note
that personally I find it primarily important to have a style enforced
in CI, but I don't have a preference as to which style is enforced.
There are quite a few options with `.clang-format`, which I'm happy to
tweak/apply as others see fit.