Shuffle `#ifdef` directives
Currently most WASIp1-vs-WASIp2 code looks like:

    #ifdef __wasip2__
    // ...
    #else
    // ...
    #endif

but when adding WASIp3 support this can get unwieldy because there's no
`#elifdef` for example. This commit reorganizes conditionally defined
code to look more like

    #if defined(__wasip1__)
    // ...
    #elif defined(__wasip2__)
    // ...
    #else
    #error "Unsupported WASI version"
    #endif

This provides a much more clear "hole" for WASIp3 support to slot into
and it additionally ensures that unknown WASI versions generate compiler
errors for various files to ensure everything is appropriately updated
when a new target is added.

This commit necessarily has quite a lot of code movement internally, but
it's not intended to have any sort of functional change.