Add a wasm32-wasip3 target
Almost all functions are stubbed out to return errors. This is all done
in preparation for eventually filling out these functions, but the goal
of this commit is to lay down the scaffolding of the target rather than
actually implement anything.

CI jobs are added to build and test this target. Most tests are expected
to fail, but all tests are still run as usual. Once tests are passing
they'll be updated in `CMakeLists.txt` to flag the test as expected to
pass instead of expected to fail.

This is a very large PR line-wise, but all the major changes with large
chunks of code are generated code and are expected to be easy to review.
The actual changes here in the source are primarily:

* Adding `#elif defined(__wasip3__)` cases
* Sometimes using the WASIp2 definitions of functions for WASIp3
* Shuffling naming in the network utilities to support both WASIp2 and
  WASIp3 which have the same types, just renamed.
* Generated bindings for WASIp3 are now vendored like WASIp2 bindings.
* Wasmtime is updated for testing to be able to run generated WASIp3
  binaries.